.archive-etf-backtest/requirements.txt
backtrader==1.9.78.123
matplotlib>=3.5.0
pyyaml>=6.0
pandas>=1.3.0
numpy>=1.20.0
seaborn>=0.11.0
yfinance>=1.0.0